Updated start scripts to not look up envrionemnt
