+++ scripts/aws_change_ownership.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
 echo "Chown application to be owned by ubuntu"
-cd /home/ubuntu/;
-chown -R ubuntu:ubuntu notifications-admin
+cd /home/notify-app/;
+chown -R notify-app:govuk-notify-applications notifications-admin
