+++ scripts/aws_install_dependencies.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
 echo "Install dependencies"
-cd /home/ubuntu/notifications-admin;
-pip3 install -r /home/ubuntu/notifications-admin/requirements.txt
+cd /home/notify-app/notifications-admin;
+pip3 install -r /home/notify-app/notifications-admin/requirements.txt
